Generate and optimize a braille-embedded ready-to-print pdf file for swell form machine or braille embosser

Opening braille device for your graphic. Just surround any graphic object between two functions brl_begin() and brl_end(). The graph will be created in PDF format in your current directory, with all text being presented using a braille font. See more details in the example section below.

brl_begin(file = "tactile.pdf", pt = 11, paper = "special",
  font = "BRL")

Arguments

file
A character string giving the filename where the image is to be saved. The default file name is "tactile.pdf" if not given.

pt
Numeric value specifying braille font size for a created file. Default is 11.

paper
Character value setting paper size of output pdf file. Default is "special".

font
Character value for font family. Available options include "BRL" for "BRAILLE1.ttf" (Braille Normal) and "SWL" for "Swell-Braille.ttf" (Swell Braille), and "TIMES" for "times.ttf" (Times New Roman). Default is set to "BRL".

Value

Nothing within the R session, but a pdf file will be created in the user's working directory.

Details

brl_begin

*Note*: This function optimizes for both 11*11 and 11*8.5 swell touch papers. When printing a generated braille pdf, just remember two things: 1. check "Fit" in the print dialog box (the shortcut key is ALT+F). 2. Check "Landscape" in the print dialog box as well. * For boxplot, you do not need to check the "Landscape," just leave it as a "Portray." However, please check the "Fit" all the time. Information: If you are not familiar with what swell form machine is, please refer to this web site. A short YouTube video will help you to understand how this function can be employed in conjunction with Swell form machine.

Examples

brl_begin(file='tactile.pdf', pt=11, paper='special', font='BRL')

#> 'font.add()' is now renamed to 'font_add()'
#> The old version still works, but consider using the new function in future code

#> 'showtext.begin()' is now renamed to 'showtext_begin()'
#> The old version still works, but consider using the new function in future code

hist(airquality$Ozone)
brl_end()

#> 'showtext.end()' is now renamed to 'showtext_end()'
#> The old version still works, but consider using the new function in future code

#> pdf 
#>   2 

# The braille ready-to-print pdf version will be created in your working directory.
# Print it out and run the swell touch paper through swell form heating machine.